5 acceptance test

+++ app/controllers/questions_controller.rb
@@ -1,37 +1,36 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
+  before_action :question, only: :show
+
   def index
     @questions = Question.all
   end
 
-  def show; end
+  def show
+    @answer = Answer.new
+  end
 
-  def new; end
-
-  def edit; end
+  def new
+    @question = current_user.questions.new
+  end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
 
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your Question was successfully created'
     else
       render :new
     end
   end
 
-  def update
-    if question.update(question_params)
-      redirect_to @question
-    else
-      render :edit
-    end
-  end
-
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if question.user.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Question was successfully deleted'
+    end
   end
 
   private
